Prepare for this template for composition
Making the answers filename match the source template avoids potential collisions. We want this to be the default so this template will work even when called alongside others pointing at the same target directory. See also [the Copier docs on pointing multiple templates to the same directory](https://copier.readthedocs.io/en/stable/configuring/#applying-multiple-templates-to-the-same-subproject).